+++ examples/next13/middleware.ts
@@ -1,20 +1,10 @@
-import { NextResponse, NextRequest } from 'next/server';
+import { createI18nMiddleware } from 'next-international/middleware';
+import { NextRequest } from 'next/server';
 
-const locales = ['en', 'fr'];
+const I18nMiddleware = createI18nMiddleware(['en', 'fr'], 'en');
 
 export function middleware(request: NextRequest) {
-  const pathname = request.nextUrl.pathname;
-  const pathnameIsMissingLocale = locales.every(
-    locale => !pathname.startsWith(`/${locale}/`) && pathname !== `/${locale}`,
-  );
-
-  if (pathnameIsMissingLocale) {
-    const acceptLanguage = request.headers.get('Accept-Language');
-    const acceptLanguageLocale = acceptLanguage?.split(',')[0].split('-')[0] ?? 'en';
-    const locale = locales.includes(acceptLanguageLocale) ? acceptLanguageLocale : 'en';
-
-    return NextResponse.redirect(new URL(`/${locale}${pathname}`, request.url));
-  }
+  return I18nMiddleware(request);
 }
 
 export const config = {
