+++ packages/next-international/middleware.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/app/middleware';
